image/requirements.txt
boto3==1.34.117
fastapi==0.111.0
mangum==0.17.0
uvicorn==0.30.0
pypdf==4.2.0
chromadb==0.5.0
langchain==0.2.1
langchain-aws==0.1.6
langchain-community==0.2.1
python-dotenv